+++ front/lib/api/skills/detection/zip/file_detection.ts
